% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{account_information}
\alias{account_information}
\title{Get account information}
\usage{
account_information(account_id, token, domain = "api-fxpractice.oanda.com")
}
\description{
Get account information
}

