% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/account_list.R
\name{account_list}
\alias{account_list}
\title{Get accounts for a user}
\usage{
account_list(token, username = NULL, accountType = c("Trade", "Practice"))
}
\arguments{
\item{token}{The personal access token. There is a link on your OANDA fxTrade account profile page titled 'Manage API Access' (My Account -> My Services -> Manage API Access). From there, you can generate a personal access token to use with the OANDA API, as well as revoke a token you may currently have.}

\item{username}{The name of the user. Note: This is only required on the sandbox, on our production systems your access token will identify you.}

\item{accountType}{The subdomain for the request is dependent on the environment you wish to obtain access tokens for. Possible choices are Practice or Trade}
}
\value{
A data frame with all the following fields: \emph{accountId}, \emph{accountName}, \emph{accountCurrency}, \emph{marginRate}
}
\description{
Get a list of accounts owned by the user
}
\examples{
account_list('my_token_1234', 'Practice')
}

