% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{close_order}
\alias{close_order}
\title{Close an order}
\usage{
close_order(oanda, order_id, account_id = NULL)
}
\arguments{
\item{order_id}{string. order ID.}

\item{account_id}{string. The account id to fetch the list of tradeable instruments for. This value will be used if it is specified.
If not specified, the value which OANDA object has inside is used as a default (default value is \code{NULL}).}
}
\description{
Close an order.
See the following link for further information:
\url{http://developer.oanda.com/rest-live/orders/#closeOrder}
}
