% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{close_position}
\alias{close_position}
\title{Close an existing position}
\usage{
close_position(oanda, instrument, account_id = NULL)
}
\arguments{
\item{instrument}{Close an open position for a specific instrument only.}

\item{account_id}{string. The account id to fetch the list of tradeable instruments for. This value will be used if it is specified.
If not specified, the value which OANDA object has inside is used as a default.}
}
\description{
Close an existing position
See the following link for further information:
\url{http://developer.oanda.com/rest-live/positions/#cloneExistingPosition}
}
