% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{close_trade}
\alias{close_trade}
\title{Close an open trade}
\usage{
close_trade(oanda, trade_id, account_id = NULL)
}
\arguments{
\item{trade_id}{string. trade ID.}

\item{account_id}{string. The account id to fetch the list of tradeable instruments for. This value will be used if it is specified.
If not specified, the value which OANDA object has inside is used as a default.}
}
\description{
Close an open trade
See the following link for further information:
\url{http://developer.oanda.com/rest-live/trades/#closeOpenTrade}
}
