% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{create_order}
\alias{create_order}
\title{Create a new order}
\usage{
create_order(oanda, instrument, units, side, type, expiry = NULL,
  price = NULL, lower_bound = NULL, upper_bound = NULL,
  stop_loss = NULL, take_profit = NULL, trailing_stop = NULL,
  accound_id = NULL)
}
\arguments{
\item{instrument}{Instrument to open the order on.}

\item{units}{The number of units to open order for.}

\item{side}{Direction of the order, either 'buy' or 'sell'.}

\item{type}{The type of the order 'limit', 'stop', 'marketIfTouched' or 'market'.}

\item{expiry}{If order type is 'limit', 'stop', or 'marketIfTouched'. The order expiration time in UTC. The value specified must be in a valid datetime format.}

\item{price}{If order type is 'limit', 'stop', or 'marketIfTouched'. The price where the order is set to trigger at.}

\item{lower_bound}{The minimum execution price.}

\item{upper_bound}{The maximum execution price.}

\item{stop_loss}{The stop loss price.}

\item{take_profit}{The take profit price.}

\item{trailing_stop}{The trailing stop distance in pips, up to one decimal place.}
}
\description{
Create a new order
See the following link for further information:
\url{http://developer.oanda.com/rest-live/orders/#createNewOrder}
}
