% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/current_prices.R
\name{current_prices}
\alias{current_prices}
\title{Get current prices}
\usage{
current_prices(token, instruments, accountType = c("Trade", "Practice"))
}
\arguments{
\item{token}{The personal access token. There is a link on your OANDA fxTrade account profile page titled 'Manage API Access' (My Account -> My Services -> Manage API Access). From there, you can generate a personal access token to use with the OANDA API, as well as revoke a token you may currently have.}

\item{instruments}{A vector of instruments to fetch prices for. Values should be one of the available instrument from the \code{instrument_list} response.}

\item{accountType}{The subdomain for the request is dependent on the environment you wish to obtain access tokens for. Possible choices are Practice or Trade}
}
\value{
A data frame with \emph{time}, \emph{bid} and \emph{ask} for all the instruments specified.
}
\description{
Fetch live prices for specified instruments that are available on the OANDA platform.
}
\examples{
current_prices('my_token_1234', c('EUR_USD', 'USD_JPY'))
}

