% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/instrument_list.R
\name{instrument_list}
\alias{instrument_list}
\title{Get an instrument list}
\usage{
instrument_list(token, accountId, accountType = c("Practice", "Trade"),
  fields = NULL, instruments = NULL)
}
\arguments{
\item{token}{The personal access token. There is a link on your OANDA fxTrade account profile page titled 'Manage API Access' (My Account -> My Services -> Manage API Access). From there, you can generate a personal access token to use with the OANDA API, as well as revoke a token you may currently have.}

\item{accountId}{The account id to fetch the list of tradeable instruments for.}

\item{accountType}{The subdomain for the request is dependent on the environment you wish to obtain access tokens for. Possible choices are Practice or Trade}

\item{fields}{A vector of instrument fields that are to be returned in the response. Please see the Value section below for a list of valid values. If the fields option is not specified, all fields will be returned}

\item{instruments}{A vector of instruments that are to be returned in the response. If the instruments option is not specified, all instruments will be returned}
}
\value{
A data frame with all or some of the following fields:\cr
\emph{instrument}: Name of the instrument. This value should be used to fetch prices and create orders and trades.\cr
\emph{displayName}: Display name for the end user.\cr
\emph{pip}: Value of 1 pip for the instrument. More on pip.\cr
\emph{maxTradeUnits}: The maximum number of units that can be traded for the instrument.\cr
\emph{precision}: The smallest unit of measurement to express the change in value between the instrument pair.\cr
\emph{maxTrailingStop}: The maximum trailing stop value (in pips) that can be set when trading the instrument.\cr
\emph{minTrailingStop}: The minimum trailing stop value (in pips) that can be set when trading the instrument.\cr
\emph{marginRate}: The margin requirement for the instrument. A 3\% margin rate will be represented as 0.03.\cr
\emph{halted}: The current trading status of the instrument. True if trading is halted, false if trading is active. \cr
}
\description{
Get a list of tradeable instruments (currency pairs, CFDs, and commodities) that are available for trading with the account specified.
}
\examples{
instrument_list('my_token_1234', 'my_account_1234', 'Practice', c('pip', 'precision'), c('AUD_CAD', 'AUD_CHF'))
}

