% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{modify_order}
\alias{modify_order}
\title{Modify an existing order}
\usage{
modify_order(oanda, order_id, units = NULL, expiry = NULL, price = NULL,
  lower_bound = NULL, upper_bound = NULL, stop_loss = NULL,
  take_profit = NULL, trailing_stop = NULL, accound_id = NULL)
}
\arguments{
\item{units}{The number of units to open order for.}

\item{expiry}{The order expiration time in UTC. The value specified must be in a valid datetime format.}

\item{price}{The price at which the order is set to trigger at.}

\item{lower_bound}{The minimum execution price.}

\item{upper_bound}{The maximum execution price.}

\item{stop_loss}{The stop loss price.}

\item{take_profit}{The take profit price.}

\item{trailing_stop}{The trailing stop distance in pips, up to one decimal place.}
}
\description{
Modify an existing order
Only the specified parameters will be modified.
All other parameters will remain unchanged.
To remove an optional parameter, set its value to 0.
See the following link for further information:
\url{http://developer.oanda.com/rest-live/orders/#modifyExistingOrder}
}
