% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{modify_trade}
\alias{modify_trade}
\title{Modify an existing trade}
\usage{
modify_trade(oanda, trade_id, stop_loss = NULL, take_profit = NULL,
  trailing_stop = NULL, accound_id = NULL)
}
\arguments{
\item{trailing_stop}{Trailing Stop distance in pips, up to one decimal place}

\item{stopLoss}{Stop Loss value}

\item{takeProfit}{Take Profit value}
}
\description{
Modify an existing trade.
Only the specified parameters will be modified.
All other parameters will remain unchanged.
To remove an optional parameter, set its value to 0.
See the following link for further information:
\url{http://developer.oanda.com/rest-live/trades/#modifyExistingTrade}
}
