% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{orders}
\alias{orders}
\title{Get orders for an account}
\usage{
orders(oanda, max_id = NULL, count = NULL, instrument = NULL,
  ids = NULL, account_id = NULL)
}
\arguments{
\item{max_id}{The server will return orders with id less than or equal to this, in descending order (for pagination).}

\item{count}{Maximum number of open orders to return. Default: 50. Max value: 500.}

\item{instrument}{Retrieve open orders for a specific instrument only. Default: all.}

\item{ids}{An URL encoded comma (%2C) separated list of orders to retrieve. Maximum number of ids: 50.
No other parameter may be specified with the ids parameter.}
}
\description{
This function will return all pending orders for an account.
See the following link for further information:
\url{http://developer.oanda.com/rest-live/orders/#getOrdersForAnAccount}
}
