% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position_list}
\alias{position_list}
\title{Get a list of open positions}
\usage{
position_list(oanda, instrument = NULL, account_id = NULL)
}
\arguments{
\item{instrument}{Retrieve open positions for a specific instrument only. Default is all( \code{NULL}).}

\item{account_id}{string. The account id to fetch the list of tradeable instruments for. This value will be used if it is specified.
If not specified, the value which OANDA object has inside is used as a default.}
}
\description{
Get a list of open positions
See the following link for further information:
\url{http://developer.oanda.com/rest-live/positions/#getListAllOpenPositions}
\url{http://developer.oanda.com/rest-live/positions/#getPositionForInstrument}
}
