% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.R
\name{price}
\alias{price}
\title{Get current prices}
\usage{
price(oanda, instruments)
}
\arguments{
\item{oanda}{OANDA class object generated by \code{\link{generate_oanda}()}}

\item{instruments}{string vector. list of instruments to fetch prices for.}
}
\description{
Fetch live prices for specified instruments that are available on the OANDA platform.
See the following link for further information:
\url{http://developer.oanda.com/rest-live/rates/#getCurrentPrices}
}
