% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{streaming_event}
\alias{streaming_event}
\title{Events Streaming}
\usage{
streaming_event(oanda, instruments, session_id = NULL, account_ids = NULL)
}
\description{
Open a streaming connection to receive real time authorized accounts' events.
}
