% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{streaming_price}
\alias{streaming_price}
\title{Rates Streaming}
\usage{
streaming_price(oanda, instruments, call_back, ignore_heartbeat = TRUE,
  session_id = NULL, account_id = NULL)
}
\arguments{
\item{call_back}{call back function which have a one argument (data.frame).}

\item{ignore_heartbeat}{boolean. whether ignore heart beat signal from OANDA Streaming API.}

\item{instruments:}{Required An URL encoded comma (%2C) separated list of instruments to fetch prices for.}

\item{session_id:}{A unique session id used to identify the rate stream connection. The value specified must be between 1 to 12 alphanumeric characters. If a request is made with a session id that matches the session id of an existing connection, the older connection will be disconnected. Please see the best practices section for usage examples.}

\item{account_id:}{The account that prices are applicable for.}
}
\description{
Open a streaming connection to receive real time market prices for specified instruments.
See the following link for further information:
\url{http://developer.oanda.com/rest-live/streaming/#ratesStreaming}
}
