% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{trade_information}
\alias{trade_information}
\title{Get information on a specific trade}
\usage{
trade_information(oanda, trade_id, accound_id = NULL)
}
\arguments{
\item{trade_id}{string. trade ID.}

\item{account_id}{string. The account id to fetch the list of tradeable instruments for. This value will be used if it is specified.
If not specified, the value which OANDA object has inside is used as a default.}
}
\description{
Get information on a specific trade.
See the following link for further information:
\url{http://developer.oanda.com/rest-live/trades/#getInformationSpecificTrade}
}
