% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_history_all}
\alias{transaction_history_all}
\title{Get full account history}
\usage{
transaction_history_all(oanda, account_id = NULL)
}
\arguments{
\item{account_id}{string. The account id to fetch the list of tradeable instruments for. This value will be used if it is specified.
If not specified, the value which OANDA object has inside is used as a default.}
}
\description{
Get full account history
See the following link for further information:
\url{http://developer.oanda.com/rest-live/transaction-history/#getFullAccountHistory}
}
