% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_information}
\alias{transaction_information}
\title{Get information for a transaction}
\usage{
transaction_information(oanda, transaction_id, account_id = NULL)
}
\arguments{
\item{account_id}{string. The account id to fetch the list of tradeable instruments for. This value will be used if it is specified.
If not specified, the value which OANDA object has inside is used as a default.}

\item{string.}{transaction ID}
}
\description{
Get information for a transaction
See the following link for further information:
\url{http://developer.oanda.com/rest-live/transaction-history/#getInformationForTransaction}
}
