% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{account_information}
\alias{account_information}
\title{Get account information}
\usage{
account_information(oanda, account_id = NULL)
}
\arguments{
\item{oanda}{OANDA class object generated by \code{\link{generate_oanda}()}}

\item{account_id}{The account id to fetch the list of tradeable instruments for. This value will be used if it is specified.
If not specified, the value which OANDA object has inside is used as a default (default value is \code{NULL}).}
}
\description{
Get account information
}
