% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{accounts}
\alias{accounts}
\title{Get accounts for a user}
\usage{
accounts(oanda)
}
\arguments{
\item{oanda}{OANDA class object generated by \code{\link{generate_oanda}()}}
}
\description{
Get a list of accounts owned by the user
}
