% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.R
\name{instrument_list}
\alias{instrument_list}
\title{Get an instrument list}
\usage{
instrument_list(oanda, instruments = NULL, account_id = NULL)
}
\description{
Get a list of tradeable instruments (currency pairs, CFDs, and commodities)
that are available for trading with the account specified.
}
