% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.R
\name{instrument_list}
\alias{instrument_list}
\title{Get an instrument list}
\usage{
instrument_list(oanda, instruments = NULL, fields = NULL,
  account_id = NULL)
}
\arguments{
\item{oanda}{OANDA class object generated by \code{\link{generate_oanda}()}}

\item{instruments}{string vector. list of instruments that are to be returned in the response.}

\item{fields}{string vector. list of instrument fields that are to be returned in the response.
The instrument field will be returned regardless of the input to this query parameter.}

\item{account_id}{string. The account id to fetch the list of tradeable instruments for. This value will be used if it is specified.
If not specified, the value which OANDA object has inside is used as a default (default value is \code{NULL}).}
}
\description{
Get a list of tradeable instruments (currency pairs, CFDs, and commodities)
that are available for trading with the account specified.
}
\details{
See the following link for further information:
\url{http://developer.oanda.com/rest-live/rates/}.
}
