% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{trade_list}
\alias{trade_list}
\title{Get a list of open trades}
\usage{
trade_list(oanda, max_id = NULL, count = NULL, instrument = NULL,
  ids = NULL, account_id = NULL)
}
\arguments{
\item{max_id}{Optional The server will return trades with id less than or equal to this, in descending order (for pagination).}

\item{count}{Optional Maximum number of open trades to return. Default: 50 Max value: 500}

\item{instrument}{Optional Retrieve open trades for a specific instrument only Default: all}

\item{ids}{Optional A (URL encoded) comma separated list of trades to retrieve. Maximum number of ids: 50. No other parameter may be specified with the ids parameter.}
}
\description{
Get a list of open trades
}
